Add local testing Makefile
Use `make test` to run test (no pun intended) for this library.

Use `make autotest` (thanks to watchexec dependency) to run test continuously
when introducing changes to either test or gem files.

Kept this ignored a long time, use it locally during development and
thought was good to have it part of the official repository.

For `autotest` to work, you will need the external `watchexec` dependency:
https://github.com/mattgreen/watchexec

[skip ci]